fixes basic test for open-api validation

+++ cr_test.sh
@@ -15,9 +15,9 @@ apiVersion: clusterregistry.k8s.io/v1alpha1
 metadata:
   name: good-cluster
 spec:
-  kubernetesApiEndpoints:
+  kubernetesAPIEndpoints:
     serverEndpoints:
-    - clientCidr: "0.0.0.0/0"
+    - clientCIDR: "0.0.0.0/0"
       serverAddress: "${SERVER}""
 
 BADCLUSTER="kind: Cluster
@@ -25,10 +25,10 @@ apiVersion: clusterregistry.k8s.io/v1alpha1
 metadata:
   name: bad-cluster
 spec:
-  kubernetesbAdiEndpoints:
+  kubernetesAPIEndpoints:
     serverEndpoints:
-      - clientCidr: "0..0/0"
-        serverbAddress: "15.640""
+      - clientCIDR: "0.0/0"
+        serverAddress: "${SERVER}""
 
 function main {
   if $(kubectl api-versions | grep "clusterregistry.k8s.io/v1alpha1" &> /dev/null); then
@@ -37,7 +37,7 @@ function main {
   fi
   kubectl create -f crd/cluster-registry.yaml
   echo "Created Cluster Registry CRD"
-  kubectl apply -f - --context ${CURRENT_CONTEXT} --validate=false <<EOF
+  kubectl apply -f - --context ${CURRENT_CONTEXT} <<EOF
 ${GOODCLUSTER}
 EOF
 if ! $(kubectl get clusters | grep "good-cluster" &> /dev/null); then
@@ -47,15 +47,21 @@ if ! $(kubectl get clusters | grep "good-cluster" &> /dev/null); then
 else
   kubectl delete cluster good-cluster
 fi
-  kubectl apply -f - --context ${CURRENT_CONTEXT} --validate=false <<EOF
+if ! $(kubectl apply -f - --context ${CURRENT_CONTEXT} <<EOF
 ${BADCLUSTER}
 EOF
+&> /dev/null); then
+echo "ERROR: bad-cluster created"
   kubectl get clusters
   kubectl delete crd/clusters.clusterregistry.k8s.io
+echo "FAIL"
+exit 1
+else
+  echo "Bad Cluster not created"
   echo "Deleted Cluster Registry CRD"
   echo "SUCCESS"
+fi
 }
-
 
 main $@
